v1.2.8
changed logic of serverside handling with how updategames works. admins now have completely authority to move whenever, even if they are a captain.  captains now have authority to move at will without a picking phase (phase is still used, but only when there are 2 actual captains) but they are limited to only their team and inbound fields to move to or from.

fixed an absolutely nasty clientside memory leak with event handler duplication. moveleft and moveright are static elements that aren't getting repainted, thus were getting events applied to them every game update. this is now being handled by adding attributes to the dom when a user clicks on a span, which the

removed picking element, as its useless now