styling patch
fixed the footer on the game page.

stripped a ton of inline font-size.

fixed the layout of joingame modal not rendering well on mobile by changing the layout of the modal.

some of the layout is a little small, maybe we should use header <hx> tags in place of some of the elements, but some will need to be custom styled in the css file.

changed the styling of the players themselves inside of game to mimick a bootstrap table.  some of this might be difficult to understand with let c=blah blah blah, but thats because i'm better with JS than i am with CS, so this is how i did it.