remove some of the styling changes from the previous commit
I kept the colors (they are cool), but the fonts I felt were ugly.  Lets stick with these current fonts, but we can adjust their size if necessary.  most of their sizes are being handled in the media queries towards the bottom of the less file if you wish to tweak them.